FPS handling in this thing was BONKERS. Mostly specified inline, not even with a variable. Changing fps from 25 to 30. This should be configurable from the CLI or from function calls but this will do for now. Fix soon.

+++ src/facerender/animate.py
@@ -194,10 +194,10 @@ class AnimateFromCoeff():
 
             try:
                 enhanced_images_gen_with_len = enhancer_generator_with_len(full_video_path, method=enhancer, bg_upsampler=background_enhancer)
-                imageio.mimsave(enhanced_path, enhanced_images_gen_with_len, fps=float(25))
+                imageio.mimsave(enhanced_path, enhanced_images_gen_with_len, fps=float(30))
             except:
                 enhanced_images_gen_with_len = enhancer_list(full_video_path, method=enhancer, bg_upsampler=background_enhancer)
-                imageio.mimsave(enhanced_path, enhanced_images_gen_with_len, fps=float(25))
+                imageio.mimsave(enhanced_path, enhanced_images_gen_with_len, fps=float(30))
             
         return_path = full_video_path
         return return_path, temp_dir
